fix: ghcr images need to publish to pact-foundation / dockerhub images to pactfoundation
GitHub container registries must contain a matching NAMESPACE to the GitHub user or org. 

This worked on my personal fork, as my GH and Docker username are the same, however pact-foundation uses a hyphen-less NAMESPACE when publishing to docker.io

https://hub.docker.com/u/pactfoundation

We therefore need to set the org name to pact-foundation when publishing to ghcr.